fix(bindings)!: honor declared track ends and unify abort error codes
Follow-ups from reviewing #2287 after it landed.

`MoqTrackProducer::finish` swallowed `Error::Closed` to paper over
`finish_at` having already declared the boundary: `finish` calls
`set_final`, which rejects a second boundary. Discarding the error hid
that, and would have hidden any other `Closed`. `track::Producer::final_sequence`
now exposes the boundary, so both bindings skip the redundant `set_final`
and say so in code.

`abort` took an `i32` and range-checked it at runtime against the `u16`
that `Error::App` actually carries. Taking a `u16` makes an out-of-range
code fail to compile and drops `MoqError::InvalidErrorCode`. The `i32`
signature was already causing drift: the Go wrapper had three `uint16`
aborts casting to `int32` next to two `int32` ones.

Go's `SubscribeJSONSnapshot` took the producer's options struct and then
ignored its `DeltaRatio`. Delta encoding is a producer-side choice the
consumer reconstructs, so subscribing takes `JSONSubscribeOptions`.

BREAKING CHANGE: `moq-ffi`'s `abort` takes a `u16` error code and
`MoqError::InvalidErrorCode` is removed. Kotlin callers pass `404u`,
matching the unsigned literals the rest of the API already needs; Swift,
Go, and Python literals are unchanged.

+++ go/wrapper/moq/publish.go
@@ -192,7 +192,7 @@ func (r *TrackRequest) Accept(info *TrackInfo) (*TrackProducer, error) {
 
 // Abort rejects the request with an application error code.
 func (r *TrackRequest) Abort(errorCode uint16) error {
-	return r.inner.Abort(int32(errorCode))
+	return r.inner.Abort(errorCode)
 }
 
 // MediaProducer writes timestamped frames into a media track.
@@ -307,7 +307,7 @@ func (t *TrackProducer) AppendDatagram(timestampUs uint64, payload []byte) (uint
 
 // Abort closes the track with an application error code.
 func (t *TrackProducer) Abort(errorCode uint16) error {
-	return t.inner.Abort(int32(errorCode))
+	return t.inner.Abort(errorCode)
 }
 
 // Consume reads directly from this producer's track. subscription tunes delivery
@@ -362,7 +362,7 @@ func (g *GroupProducer) Finish() error {
 
 // Abort terminates the group with an application error code.
 func (g *GroupProducer) Abort(errorCode uint16) error {
-	return g.inner.Abort(int32(errorCode))
+	return g.inner.Abort(errorCode)
 }
 
 // TrackDynamic yields uncached groups requested by fetch consumers.
@@ -409,7 +409,7 @@ func (r *GroupRequest) Accept() (*GroupProducer, error) {
 }
 
 // Abort rejects the fetch with an application error code.
-func (r *GroupRequest) Abort(errorCode int32) error {
+func (r *GroupRequest) Abort(errorCode uint16) error {
 	return r.inner.Abort(errorCode)
 }
