+++ swift/Sources/Moq/Dynamic.swift
@@ -27,7 +27,7 @@ public final class TrackRequest: Sendable {
     }
 
     /// Reject the request with an application error code, failing the subscriber.
-    public func abort(errorCode: Int32) throws {
+    public func abort(errorCode: UInt16) throws {
         try ffi.abort(errorCode: errorCode)
     }
 }
@@ -56,7 +56,7 @@ public final class GroupRequest: Sendable {
     }
 
     /// Reject the fetch with an application error code.
-    public func abort(errorCode: Int32) throws {
+    public func abort(errorCode: UInt16) throws {
         try ffi.abort(errorCode: errorCode)
     }
 }
